iter3 completed
